SA-16913 Setup data for FusionDirectory before replication starts
init.sh, which setup data for FusionDirectory, assumes that it is
executed in the clean state in which data for FusionDirectory is not
stored in its database. And it also means that init.sh assumes that it
is executed before replication starts, since the data for
FusionDirectory can be created by replication. As a result, init.sh
can fail if init.sh is executed after replication starts. Therefore,
this commit changes the location at which init.sh is executed so that
it is executed before replication starts.